% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_cd_municipio}
\alias{.extract_cd_municipio}
\title{Extrai o código do município do código da unidade gestora. 
Essa estração é feita através da coleta dos últimos 3 digitos da unidade gestora.}
\usage{
.extract_cd_municipio(df, cd_u_gestora)
}
\arguments{
\item{df}{Dataframe sem o código do município}

\item{cd_u_gestora}{código da unidade gestora}
}
\value{
Dataframe com o código do município
}
\description{
Extrai o código do município do código da unidade gestora. 
Essa estração é feita através da coleta dos últimos 3 digitos da unidade gestora.
}
