% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_hash_id}
\alias{.generate_hash_id}
\title{Gera um identificador único para cada registro do dataframe.
Usa o algoritmo md5 para gerar um hash a partir da concatenação das colunas
que devem ser chave primária do dataframe passado como parâmetro.}
\usage{
.generate_hash_id(df, colunas, id_coluna)
}
\arguments{
\item{df}{Dataframe sem identificador único}

\item{colunas}{array com o nome das colunas que são chave primária do dataframe}

\item{id_coluna}{Nome da coluna do identificador}
}
\value{
Dataframe com identificador único
}
\description{
Gera um identificador único para cada registro do dataframe.
Usa o algoritmo md5 para gerar um hash a partir da concatenação das colunas
que devem ser chave primária do dataframe passado como parâmetro.
}
