% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.substrRight}
\alias{.substrRight}
\title{Extrai os últimos n caracteres de uma string.}
\usage{
.substrRight(x, n)
}
\arguments{
\item{x}{string com todos os caracteres}

\item{n}{numero de caracteres que serão extraídos}
}
\value{
os últimos n caracteres
}
\description{
Extrai os últimos n caracteres de uma string.
}
