% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fncols}
\alias{fncols}
\title{Adiciona uma nova coluna caso ela não exista}
\usage{
fncols(data, cname)
}
\arguments{
\item{data}{dataframe principal}

\item{cname}{colunas que serão verificadas}
}
\value{
dataframe com as novas colunas
}
\description{
Adiciona uma nova coluna caso ela não exista
}
