% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{format_contrato_tce_rs}
\alias{format_contrato_tce_rs}
\title{Processa dataframe dos contratos}
\usage{
format_contrato_tce_rs(contratos_transformados_rs_df)
}
\arguments{
\item{contratos_df}{Dataframe contendo informações dos contratos}
}
\value{
Dataframe contendo informações dos contratos processados
}
\description{
Manipula a tabela para se adequar aos outros estados
}
