% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{format_empenhos_tce_rs}
\alias{format_empenhos_tce_rs}
\title{Processa dataframe dos empenhos}
\usage{
format_empenhos_tce_rs(empenhos_pagamentos_transformados_rs_df)
}
\arguments{
\item{empenhos_pagamentos_transformados_rs_df}{Dataframe contendo informações de empenhos e pagamentos do RS}
}
\value{
Dataframe contendo informações dos empenhos
}
\description{
Manipula a tabela para se adequar aos outros estados
}
