% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{format_licitacao_tce_rs}
\alias{format_licitacao_tce_rs}
\title{Processa dataframe de licitações}
\usage{
format_licitacao_tce_rs(licitacoes_transformadas_rs_df)
}
\arguments{
\item{licitacoes_df}{Dataframe contendo informações das licitações}
}
\value{
Dataframe contendo informações das licitações processados
}
\description{
Manipula a tabela para se adequar aos outros estados
}
