% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{format_pagamentos_tce_rs}
\alias{format_pagamentos_tce_rs}
\title{Processa dataframe de pagamentos}
\usage{
format_pagamentos_tce_rs(empenhos_pagamentos_transformados_rs_df)
}
\arguments{
\item{empenhos_pagamentos_transformados_rs_df}{Dataframe contendo informações de empenhos e pagamentos do RS}
}
\value{
Dataframe contendo informações dos pagamentos
}
\description{
Manipula a tabela para se adequar aos outros estados
}
