% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_contrato_id}
\alias{generate_contrato_id}
\title{Gera um identificador único para cada contrato}
\usage{
generate_contrato_id(contratos_df)
}
\arguments{
\item{contratos_df}{Dataframe contendo informações sobre os contratos}
}
\value{
Dataframe contendo informações sobre os contratos e seus ids
}
\description{
Gera um identificador único para cada contrato
}
\examples{
contratos_dt <- generate_contrato_id(contrato_df)
}
