% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_contrato_tramita_id}
\alias{generate_contrato_tramita_id}
\title{Gera um identificador único para cada contrato de 2020}
\usage{
generate_contrato_tramita_id(contratos_df)
}
\arguments{
\item{contratos_df}{Dataframe contendo informações sobre os contratos de 2020}
}
\value{
Dataframe contendo informações sobre os contratos de 2020 e seus ids
}
\description{
Gera um identificador único para cada contrato de 2020
}
\examples{
contratos_dt <- generate_contrato_tramita_id(contrato_df)
}
