% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_empenho_id}
\alias{generate_empenho_id}
\title{Gera um identificador único para cada empenho}
\usage{
generate_empenho_id(empenho_df)
}
\arguments{
\item{empenhos_df}{Dataframe contendo informações sobre os empenhos}
}
\value{
Dataframe contendo informações sobre os empenhos e seus ids
}
\description{
Gera um identificador único para cada empenho
}
\examples{
empenhos_dt <- generate_empenho_id(empenho_df)
}
