% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{generate_empenhos_pagamento_tce_rs_id}
\alias{generate_empenhos_pagamento_tce_rs_id}
\title{Gera um identificador único para empenhos e licitações do RS}
\usage{
generate_empenhos_pagamento_tce_rs_id(empenhos_pagamento_tce_rs, file_empenho)
}
\arguments{
\item{empenhos_pagamento_tce_rs}{Dataframe contendo informações sobre empenhos e licitações}
}
\value{
Dataframe contendo informações sobre empenhos e licitações do RS e seus ids
}
\description{
Gera um identificador único para empenhos e licitações do RS
}
\examples{
empenhos_pagamento_tce_rs <- generate_empenhos_pagamento_tce_rs_id(empenhos_pagamento_tce_rs)
}
