% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_estorno_pagamento_id}
\alias{generate_estorno_pagamento_id}
\title{Gera um identificador único para cada estorno de pagamento}
\usage{
generate_estorno_pagamento_id(estorno_pagamento_df)
}
\arguments{
\item{estorno_pagamento_df}{Dataframe contendo informações sobre os estornos de pagamentos}
}
\value{
Dataframe contendo informações sobre os estornos de pagamentos e seus ids
}
\description{
Gera um identificador único para cada estorno de pagamento
}
\examples{
estorno_pagamento_dt <- generate_estorno_pagamento_id(estorno_pagamento_df)
}
