% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_licitacao_id}
\alias{generate_licitacao_id}
\title{Gera um identificador único para cada licitação}
\usage{
generate_licitacao_id(licitacoes_df)
}
\arguments{
\item{licitacoes_df}{Dataframe contendo informações sobre as licitações}
}
\value{
Dataframe contendo informações sobre as licitações e seus ids
}
\description{
Gera um identificador único para cada licitação
}
\examples{
licitacoes_dt <- generate_licitacao_id(licitacoes_df)
}
