% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{generate_licitacao_tce_rs_id}
\alias{generate_licitacao_tce_rs_id}
\title{Gera um identificador único para cada licitação do RS}
\usage{
generate_licitacao_tce_rs_id(licitacoes_tce_rs_dt)
}
\arguments{
\item{licitacoes_df}{Dataframe contendo informações sobre as licitações}
}
\value{
Dataframe contendo informações sobre as licitações e seus ids
}
\description{
Gera um identificador único para cada licitação do RS
}
\examples{
licitacoes_tce_rs_dt <- generate_licitacao_tce_rs_id(licitacoes_tce_rs_dt)
}
