% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_licitacao_tramita_id}
\alias{generate_licitacao_tramita_id}
\title{Gera um identificador único para cada licitação do tramita}
\usage{
generate_licitacao_tramita_id(licitacoes_tramita_df)
}
\arguments{
\item{licitacoes_tramita_df}{Dataframe contendo informações sobre as licitações}
}
\value{
Dataframe contendo informações sobre as licitações contidas no Tramita e seus ids
}
\description{
Gera um identificador único para cada licitação do tramita
}
\examples{
licitacoes_tramita_dt <- generate_licitacao_tramita_id(licitacoes_tramita_df)
}
