% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_pagamento_id}
\alias{generate_pagamento_id}
\title{Gera um identificador único para cada pagamento}
\usage{
generate_pagamento_id(pagamentos_df)
}
\arguments{
\item{pagamentos_df}{Dataframe contendo informações sobre os pagamentos}
}
\value{
Dataframe contendo informações sobre os pagamentos e seus ids
}
\description{
Gera um identificador único para cada pagamento
}
\examples{
pagamentos_dt <- generate_pagamento_id(pagamento_df)
}
