% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{generate_proposta_id}
\alias{generate_proposta_id}
\title{Gera um identificador único para cada proposta}
\usage{
generate_proposta_id(propostas_df)
}
\arguments{
\item{propostas_df}{Dataframe contendo informações sobre as propostas}
}
\value{
Dataframe contendo informações sobre as propostas e seus ids
}
\description{
Gera um identificador único para cada proposta
}
\examples{
propostas_dt <- generate_proposta_id(propostas_df)
}
