% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_codigo_localidades_ibge}
\alias{process_codigo_localidades_ibge}
\title{Processa dataframe de codigos das localidades do IBGE}
\usage{
process_codigo_localidades_ibge(codigo_localidades_ibge_df)
}
\arguments{
\item{process_codigo_localidades_df}{Dataframe contendo informações
dos codigos das localidades do IBGE}
}
\value{
Dataframe contendo informações dos codigos das localidades do IBGE
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
