% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_contrato}
\alias{process_contrato}
\title{Processa dataframe de contratos}
\usage{
process_contrato(contratos_df)
}
\arguments{
\item{contratos_df}{Dataframe contendo informações dos contratos}
}
\value{
Dataframe contendo informações dos contratos processados
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
