% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_contrato_mutado}
\alias{process_contrato_mutado}
\title{Processa dataframe de contratos mutados}
\usage{
process_contrato_mutado(contratos_mutados_df)
}
\arguments{
\item{contratos_mutados_df}{Dataframe contendo informações dos contratos mutados}
}
\value{
Dataframe contendo informações dos contratos mutados processados
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
