% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{process_contrato_tce_rs}
\alias{process_contrato_tce_rs}
\title{Processa dataframe de contratos}
\usage{
process_contrato_tce_rs(contratos_df)
}
\arguments{
\item{contratos_df}{Dataframe contendo informações dos contratos}
}
\value{
Dataframe contendo informações dos contratos processados
}
\description{
Manipula a tabela pra forma que será utilizada no banco
}
