% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_contrato_tramita}
\alias{process_contrato_tramita}
\title{Processa dataframe de contratos de 2020}
\usage{
process_contrato_tramita(contratos_tramita_df)
}
\arguments{
\item{contratos_df}{Dataframe contendo informações dos contratos de 2020}
}
\value{
Dataframe contendo informações dos contratos processados de 2020
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
