% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_distritos_ibge}
\alias{process_distritos_ibge}
\title{Processa dataframe de distritos}
\usage{
process_distritos_ibge(distritos_ibge_df)
}
\arguments{
\item{distritos_ibge_df}{Dataframe contendo informações dos contratos mutados}
}
\value{
Dataframe contendo informações dos distritos  processados
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
