% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_empenho}
\alias{process_empenho}
\title{Processa dataframe de empenhos}
\usage{
process_empenho(empenhos_df)
}
\arguments{
\item{empenhos_df}{Dataframe contendo informações dos empenhos}
}
\value{
Dataframe contendo informações dos empenhos processados
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
