% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{process_empenhos_pagamento_tce_rs}
\alias{process_empenhos_pagamento_tce_rs}
\title{Processa dataframe de empenhos e pagamentos do RS}
\usage{
process_empenhos_pagamento_tce_rs(empenhos_pagamento_tce_rs, file_empenho)
}
\arguments{
\item{empenhos_pagamento_ce_rs}{Dataframe contendo informações de empenhos e pagamentos do RS}
}
\value{
Dataframe contendo informações dos empenhos e pagamentos
}
\description{
Manipula a tabela para a adição de identificadores e para adequação
             de colunas.
}
