% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_estorno_pagamento}
\alias{process_estorno_pagamento}
\title{Processa dataframe dos estornos de pagamentos}
\usage{
process_estorno_pagamento(estorno_pagamento_df)
}
\arguments{
\item{estorno_pagamento_df}{Dataframe contendo informações dos estornos de pagamentos}
}
\value{
Dataframe contendo informações dos estornos de pagamentos processados
}
\description{
Manipula tabela que será utilizada no banco
}
