% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_licitacao}
\alias{process_licitacao}
\title{Processa dataframe de licitações}
\usage{
process_licitacao(licitacoes_df)
}
\arguments{
\item{licitacoes_df}{Dataframe contendo informações das licitações}
}
\value{
Dataframe contendo informações das licitações processados
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
