% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tce_rs_opendata.R
\name{process_licitacao_tce_rs}
\alias{process_licitacao_tce_rs}
\title{Processa dataframe de licitações}
\usage{
process_licitacao_tce_rs(licitacoes_df)
}
\arguments{
\item{licitacoes_df}{Dataframe contendo informações das licitações}
}
\value{
Dataframe contendo informações das licitações processados
}
\description{
Manipula a tabela pra forma que será utilizada no banco
}
