% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_licitacao_tramita}
\alias{process_licitacao_tramita}
\title{Processa dataframe das licitações contidas no tramita}
\usage{
process_licitacao_tramita(licitacoes_tramita_df)
}
\arguments{
\item{licitacoes_df}{Dataframe contendo informações das licitações}
}
\value{
Dataframe contendo informações das licitações do tramita processados
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
