% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_municipio}
\alias{process_municipio}
\title{Processa datagrame de municipios}
\usage{
process_municipio(municipios_df)
}
\arguments{
\item{municipios_df}{Dataframe contendo informações dos municipios}
}
\value{
Dataframe contendo informações dos municipios processados
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
