% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_pagamento}
\alias{process_pagamento}
\title{Processa dataframe dos pagamentos}
\usage{
process_pagamento(pagamentos_df)
}
\arguments{
\item{pagamentos_df}{Dataframe contendo informações dos pagamentos}
}
\value{
Dataframe contendo informações os pagamentos processados
}
\description{
Manipula tabela que será utilizada no banco
}
