% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_participante}
\alias{process_participante}
\title{Processa dataframe dos participantes}
\usage{
process_participante(participantes_df)
}
\arguments{
\item{participantes_df}{Dataframe contendo informações dos participantes}
}
\value{
Dataframe contendo informações os participantes processados
}
\description{
Manipula tabela que será utilizada no banco
}
