% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R
\name{process_proposta}
\alias{process_proposta}
\title{Processa dataframe de propostas}
\usage{
process_proposta(propostas_df)
}
\arguments{
\item{propostas_df}{Dataframe contendo informações das propostas}
}
\value{
Dataframe contendo informações das propostas processadss
}
\description{
Manipula tabela pra forma que será utilizada no banco
}
