% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres.R, R/tce_rs_opendata.R
\name{generate_contrato_id}
\alias{generate_contrato_id}
\alias{generate_contrato_tce_rs_id}
\title{Gera um identificador único para cada contrato}
\usage{
generate_contrato_id(contratos_df)

generate_contrato_tce_rs_id(contratos_tce_rs_dt)
}
\arguments{
\item{contratos_df}{Dataframe contendo informações sobre os contratos}
}
\value{
Dataframe contendo informações sobre os contratos e seus ids

Dataframe contendo informações sobre os contratos e seus ids
}
\description{
Gera um identificador único para cada contrato

Gera um identificador único para cada contrato do RS
}
\examples{
contratos_dt <- generate_contrato_id(contrato_df)
contratos_tce_rs_dt <- generate_contrato_tce_rs_id(contrato_df)
}
